refactor: move player profile update slice into core application
