+++ backend/src/api/interface/controller/v1/players_controller.py
@@ -12,11 +12,11 @@ from api.middleware.exception_mapper import map_exceptions
 from auth.dependencies import authorize_player_access, require_user
 from core.application.models import PlayerProfile
 from core.application.use_cases.get_player_profile_usecase import GetPlayerProfileUseCase
-from fastapi import APIRouter, Depends, HTTPException, status
-from persistence.application.use_cases import (
+from core.application.use_cases.update_player_profile_usecase import (
     PlayerUpdate,
     UpdatePlayerProfileUseCase,
 )
+from fastapi import APIRouter, Depends, HTTPException, status
 
 router = APIRouter()
