+++ backend/tests/test_update_player_profile_use_case.py
@@ -1,19 +1,19 @@
 from dataclasses import dataclass
 
 import pytest
+from core.application.use_cases.update_player_profile_usecase import (
+    InvalidNationalityError,
+    InvalidPlayerUpdateDataError,
+    InvalidProfileImageError,
+    PlayerUpdate,
+    UpdatePlayerProfileUseCase,
+)
 from persistence.application.ports.player_profile_port import (
     InvalidNationalityError as RepositoryInvalidNationalityError,
 )
 from persistence.application.ports.player_profile_port import (
     PenaInfoResult,
     PlayerProfileResult,
-)
-from persistence.application.use_cases.update_player_profile_usecase import (
-    InvalidNationalityError,
-    InvalidPlayerUpdateDataError,
-    InvalidProfileImageError,
-    PlayerUpdate,
-    UpdatePlayerProfileUseCase,
 )
